create UI methods and modify service methods to working status
